% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimePointsPenalized.R
\name{fitTimePointsPenalized.cv}
\alias{fitTimePointsPenalized.cv}
\title{CV with penalized differences between adjacent time points coefficients}
\usage{
fitTimePointsPenalized.cv(
  y0,
  x0,
  FollowUp,
  lam1V,
  gamma,
  tV,
  standardize = TRUE,
  Clinical0 = data.frame(case_control0 = y0),
  startWithGlmnet = FALSE,
  folds
)
}
\arguments{
\item{y0}{case/control vector (no time iformation - "naive" approach)}

\item{x0}{gene expression matrix (rows-samples by columns-genes)}

\item{FollowUp}{follow-up times (recurrence time for recurrences and follow-up for patients with no recurrences)}

\item{lam1V}{array of lasso penalty prefactor}

\item{gamma}{prefactor of the second penalty term - differences between adjacent time points coefficients}

\item{tV}{array of time points}

\item{standardize}{TRUE/FALSE standardization of the x0 columns (zero mean, unit variance)}

\item{Clinical0}{dataframe with clinical information (same order as rows of x0)}
}
\description{
CV with penalized differences between adjacent time points coefficients
}
