% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimePointsPenalized.R
\name{fitTimePointsPenalized}
\alias{fitTimePointsPenalized}
\title{Fit lasso with penalized differences between adjacent time points coefficients}
\usage{
fitTimePointsPenalized(
  y0,
  x0,
  FollowUp,
  lam1V,
  gamma,
  alpha,
  tV,
  Clinical0 = data.frame(case_control0 = y0),
  startWithGlmnet = FALSE
)
}
\arguments{
\item{y0}{case/control vector (no time iformation - "naive" approach)}

\item{x0}{gene expression matrix (rows-samples by columns-genes)}

\item{FollowUp}{follow-up times (recurrence time for recurrences and follow-up for patients with no recurrences)}

\item{lam1V}{array of lasso penalty prefactor}

\item{gamma}{prefactor of the second penalty term - differences between adjacent time points coefficients}

\item{alpha}{glmnet parameter}

\item{tV}{array of time points}

\item{Clinical0}{dataframe with clinical information (same order as rows of x0)}
}
\description{
Fit lasso with penalized differences between adjacent time points coefficients
}
