% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimePointsPenalized.R
\docType{package}
\name{TimePointsPenalized}
\alias{TimePointsPenalized}
\title{TimePointsPenalized: Lasso with penalized differences between adjacent time points}
\section{Mypackage functions}{

A package to fit lasso with penalized differences between adjacent time points coefficients.
}

